Added Admin Functionality and updated some code

+++ src/app/show-task/Task.jsx
@@ -1,13 +1,35 @@
 import UserContext from "@/context/userContext";
 import { deleteTask } from "@/services/todoSevices";
+import { data } from "autoprefixer";
 import React, { useContext } from "react";
 import { RxCross1 } from "react-icons/rx";
+import Swal from "sweetalert2";
 
 function Task({ task, handleDeleteTaskParent }) {
   const { user } = useContext(UserContext);
 
+  const date = new Date();
+  console.log("Date", date.toLocaleTimeString());
+
   const handleDeleteTask = () => {
-    handleDeleteTaskParent(task._id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        handleDeleteTaskParent(task._id);
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success",
+        });
+      }
+    });
   };
 
   return (
@@ -24,13 +46,17 @@ function Task({ task, handleDeleteTaskParent }) {
         <p className="font-normal">{task.content}</p>
         <div className="flex justify-between mt-3">
           <p className="text-left">
-            Status:{" "}
+            Status:
             <span className="font-bold">{task.status.toUpperCase()}</span>
           </p>
           <p className="text-right">
             Author: <span className="font-bold">{user?.name}</span>
           </p>
         </div>
+        <p className="text-left">
+          Added on : {date.toLocaleDateString()} at
+          {"  " + date.toLocaleTimeString()}
+        </p>
       </div>
     </div>
   );
